Move legacy dimension declaration to JNI library
JNI library is the only one still using the older API. All other
component use the model.
