+++ subprojects/platform-base/src/main/java/dev/nokee/platform/base/internal/BaseComponent.java
@@ -19,12 +19,8 @@ import dev.nokee.model.internal.core.ModelNode;
 import dev.nokee.model.internal.core.ModelNodeAware;
 import dev.nokee.model.internal.core.ModelNodeContext;
 import dev.nokee.platform.base.*;
-import dev.nokee.runtime.core.CoordinateSet;
-import dev.nokee.runtime.core.CoordinateSpace;
-import dev.nokee.utils.Cast;
 import lombok.Getter;
 import org.gradle.api.model.ObjectFactory;
-import org.gradle.api.provider.ListProperty;
 import org.gradle.api.provider.Property;
 import org.gradle.api.provider.Provider;
 import org.gradle.api.provider.SetProperty;
@@ -32,32 +28,11 @@ import org.gradle.api.provider.SetProperty;
 public abstract class BaseComponent<T extends Variant> implements Component, ModelNodeAware {
 	private final ModelNode node = ModelNodeContext.getCurrentModelNode();
 	@Getter private final ComponentIdentifier identifier;
-
-	// TODO: We may want to model this as a DimensionRegistry for more richness than a plain set
-	private final ListProperty<CoordinateSet<?>> dimensions;
-	private final Property<CoordinateSpace> finalSpace;
-
 	@Getter private final Property<String> baseName;
 
 	protected BaseComponent(ComponentIdentifier identifier, ObjectFactory objects) {
 		this.identifier = identifier;
-		this.dimensions = Cast.uncheckedCastBecauseOfTypeErasure(objects.listProperty(CoordinateSet.class));
-		this.finalSpace = objects.property(CoordinateSpace.class);
 		this.baseName = objects.property(String.class);
-
-		getDimensions().finalizeValueOnRead();
-
-		getFinalSpace().convention(getDimensions().map(CoordinateSpace::cartesianProduct));
-		getFinalSpace().disallowChanges();
-		getFinalSpace().finalizeValueOnRead();
-	}
-
-	public ListProperty<CoordinateSet<?>> getDimensions() {
-		return dimensions;
-	}
-
-	public Property<CoordinateSpace> getFinalSpace() {
-		return finalSpace;
 	}
 
 	public abstract Provider<T> getDevelopmentVariant();
